src/requirements.txt
# setuptools>=v67.7.1
django>=4.2.6
celery>=5.3.4
django-storages>=1.14.2
django-session-timeout>=0.1.0
django-cors-headers>=4.3.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.1
django-rest-durin>=1.1.0
djangorestframework>=3.14.0
django-db-geventpool>=4.0.1
psycopg2-binary>=2.9.9
requests>=2.31.0
gunicorn>=21.2.0
gevent>=23.9.1
drf-yasg>=1.19.4
pytz>=2023.3.post1
sentry-sdk>=1.32.0
boto3>=1.28.63